Crate libaoc

This a library containing functions, structs, enums, traits and methods for common little problems while solving the Advent of Code.

This library has a trait for converting Iterators, a struct and an enum for keeping track of a Position and a Direction, a trait for calculating the manhatten-distance, a trait implemented on all integers, that allow to get the absolute value of that integer, and a trait for quickly sorting a tuple in ascending or descending order.

Also supports reading tekst from a file into a String, or Vec, however this is a feature of this library, and is considered unstable.

Modules

absolute

convert

movement

Macros

add
Adds all items in a tuple.

apply
Applies any given operator to any given tuple.

div
Divides all items in a tuple. Panics if divided by 0.

mul
Multiplies all items in a tuple.

noop
This macro is used to generate noop functions. Works only on tuples, but is great to use in functional programming when you just need to pass a noop function into another function.

rem
'Remainders' all items in a tuple.

sub
Subtracts all items in a tuple.

Traits

MinMax
2 functions to 'sort' a tuple. minmax returns the tuple in ascending order, maxmin in descending order.

Functions

sort_biggest
Returns a tuple, sorted by the max value.

sort_smallest
Returns a tuple, sorted by the min value.