Trait libaoc::convert::Convert

pub trait Convert<T, U, I> where
    U: From<T>,
    I: Iterator<Item = T>, {
    type Iterable: Iterator<Item = U>;
    fn convert(self) -> Vec<U>;

    fn convert_into_slice(self, slice: &mut [U]) -> usize;

    fn convert_iter(self) -> Self::Iterable;
}

This trait allows to convert a stream with items of type T into a stream or collection with items of type U.

Examples

extern crate libaoc;
 
use libaoc::convert::Convert;
use libaoc::movement::Position;
 
fn main() {
    let tuple1 = (0, 0);
    let tuple2 = (1, 1);
    let tuple3 = (2, 2);

    let tuples = vec![tuple1, tuple2, tuple3];

    let Positions: Vec<Position<usize>> = tuples.into_iter().convert();

    assert_eq!(vec![Position::new(0, 0), Position::new(1, 1), Position::new(2, 2)], Positions);
 
    let tups = vec![(4, 4), (5, 5), (3, 4)];
 
    let mut convert_iter = tups.into_iter().rev().convert_iter();
 
    assert_eq!(Some(Position::new(3, 4)), convert_iter.next());
}

Associated Types

type Iterable: Iterator<Item = U>

The Iterator that gets returned from convert_iter()

Required Methods

fn convert(self) -> Vec<U>

Returns a vector of all completed conversions.

fn convert_into_slice(self, slice: &mut [U]) -> usize

Converts the stream, and writes the items into slice. Returns how many elements where written to the slice.

Examples

extern crate libaoc;
 
use libaoc::convert::Convert;
 
fn main() {
    let chars = vec![97, 98, 99, 100, 101];
    let mut slice: [char; 5] = ['-'; 5];
 
    let written = chars.into_iter().convert_into_slice(&mut slice);
     
    assert_eq!(['a', 'b', 'c', 'd', 'e'], slice);
    assert_eq!(5, written);
}

fn convert_iter(self) -> Self::Iterable

Returns an iterator that performs the conversions.