Module libaoc::movement

Structs

Position
A position. Great to use in maps or graphs.

Enums

Direction
An enum to represent a direction. Is great to use in maps, or when 'following' some kind of line. Also supports an init variant. When the current variant is init, using turn_right returns the variant right. When the current variant is init, using turn_left returns the variant left.

Traits

ManhattenDst
Returns the manhatten distance of any Position with type N. A position is either a tuple, or the struct Position. the manhatten distance is the sum of the absolute values of a coordinate.