Struct libaoc::movement::Position

pub struct Position<N> { /* fields omitted */ }

A position. Great to use in maps or graphs.

Examples

extern crate libaoc;
use libaoc::movement::{Position, ManhattenDst};
use libaoc::absolute::Absolute;
fn main() {
    let tup = (-10i32, 21i32);
    let pos = Position::new(-10i32, 21i32);
    assert_eq!(Position::from(tup), pos);

    let tuple = (10u16, 1u16);
    let position = Position::new(10u16, 1u16);

    assert_eq!(Position::from(tuple), position);

    let p1 = Position::new(10, 20);
    let p2 = Position::new(20, 30);

    assert_eq!(Position::new(10, 10), p2 - p1);

    let rp1 = &Position::new(10, 20);
    let otherp2 = Position::new(-20, 30);

    let p3 = otherp2 - rp1;
    assert_eq!(Position::new(-30, 10), p3);

    assert_eq!(Position::new(30, 10), p3.abs());

    assert_eq!(Position::new(-10, 40), p3 + p2);
}

Methods

impl<N> Position<N> where
N: AddAssign<N> + SubAssign<N>,

pub fn new(x: N, y: N) -> Position<N>

Returns a new Position.

pub fn change(&mut self, direction: &Direction, steps: N)

Changes the position with steps based on the direction. If the direction is facing down, y is incremented, if the direction if facing up, y is decremented. If the direction is Direction::Init, no update is made.

Examples

extern crate libaoc;
use libaoc::movement::{Position, Direction};
fn main() {
    let mut pos = Position::new(0, 0);
    let dir = Direction::init_up();

    pos.change(&dir, 1);

    let otherpos = Position::new(0, -1);
    assert_eq!(pos, otherpos);
}

pub fn rev_change(&mut self, direction: &Direction, steps: N)

Same as change, but now increments y when facing upwards, and decrements y when facing downwards.

Examples

extern crate libaoc;
use libaoc::movement::{Direction, Position};
fn main() {
    let mut pos = Position::new(0, 0);
    let dir = Direction::init_up();

    pos.rev_change(&dir, 1);

    let otherpos = Position::new(0, 1);
    assert_eq!(pos, otherpos);
}

pub fn is_adjecent<'a, 'b>(&'a self, other: &'b Position<N>) -> bool where
N: Sub<Output = N> + Clone + From<i8> + PartialEq + Absolute,

Check whether self and other are adjecent. That is, if the absolute x value and the absolute y value after subtracting self from other is either (1, 0), (0, 1) or (1, 1).

Examples

extern crate libaoc;
use libaoc::movement::Position;
use libaoc::absolute::Absolute;
fn main() {
    let pos1 = Position::new(5, 6);
    let pos2 = Position::new(6, 7);

    assert_eq!(true, pos1.is_adjecent(&pos2));
    assert_eq!(true, pos2.is_adjecent(&pos1));

    let pos3 = Position::new(-1, 0);
    let pos4 = Position::new(0, 0);

    assert_eq!(true, pos3.is_adjecent(&pos4));
    assert_eq!(true, pos4.is_adjecent(&pos3));

    assert_eq!(false, pos4.is_adjecent(&pos4));
    assert_eq!(false, pos3.is_adjecent(&pos1));
}

pub fn increment_y(&mut self, steps: N)

Adds steps to y.

Examples

extern crate libaoc;
use libaoc::movement::Position;
fn main() {
    let mut pos = Position::new(0, 0);
    pos.increment_y(10);
    assert_eq!(Position::new(0, 10), pos);
}

pub fn decrement_y(&mut self, steps: N)

Subtracts steps from y.

pub fn increment_x(&mut self, steps: N)

Adds steps to x.

pub fn decrement_x(&mut self, steps: N)

Subtracts steps from x.

pub fn get_ref_x(&self) -> &N

Returns a reference to the current x value.

pub fn get_ref_y(&self) -> &N

Returns a reference to the current y value.

pub fn get_ref(&self) -> (&N, &N)

Returns a tuple of &x, &y.

pub fn to_tuple(&self) -> (N, N) where
N: Clone,

Clones x and y into a tuple.