Trait libaoc::TryConvert

pub trait TryConvert<N, S, I> where
    N: FromStr,
    S: AsRef<str>,
    I: Iterator<Item = S>, {
    fn try_convert(self) -> Result<Vec<N>, <N as FromStr>::Err>;

    unsafe fn unsafe_convert(self) -> Vec<N>
    where
        <N as FromStr>::Err: Debug;

    fn try_convert_iter(
        self
    ) -> Map<I, fn(_: S) -> Result<N, <N as FromStr>::Err>>;

    unsafe fn unsafe_convert_iter(self) -> Map<I, fn(_: S) -> N>
    where
        <N as FromStr>::Err: Debug;
}

Takes any Iterator, where the items implement AsRef. Returns a Vec, where N implements FromStr. Returns an Error if an error occured.

Examples

extern crate libaoc;
use libaoc::{TryConvert};

fn main () {
    let s = "1, 2, 3, 4, 5";
    assert_eq!(vec![1, 2, 3, 4, 5], s.split(", ").try_convert().unwrap());

    let s = String::from("1\n2\n3\n4\n5\n6");
    assert_eq!(vec![1, 2, 3, 4, 5, 6], s.lines().try_convert().unwrap());

    let my_str = "1, 2, 3,4, 5";
    let mut iter = my_str.split(", ").try_convert_iter();

    assert_eq!(Some(Ok(1)), iter.next());
    assert_eq!(Some(Ok(2)), iter.next());
 
    //the next would be 3, but that gives an error. assert_NE.
    assert_ne!(Some(Ok(3)), iter.next());
 
    //due to the split call, 4 isn't part of the stream, so next one is 5.
    assert_eq!(Some(Ok(5)), iter.next());
}

Required Methods

fn try_convert(self) -> Result<Vec<N>, <N as FromStr>::Err>

On succes, returns a vector of all completed conversions. When an error occures, returns an error instead.

unsafe fn unsafe_convert(self) -> Vec<N> where
<N as FromStr>::Err: Debug,

On succes, returns a vector of all completed conversions.

Panic

Panics when an error occures.

fn try_convert_iter(self) -> Map<I, fn(_: S) -> Result<N, <N as FromStr>::Err>>

Returns an iterator over the converted items. Returns an error if an item can not be converted. Continue's after the error.

unsafe fn unsafe_convert_iter(self) -> Map<I, fn(_: S) -> N> where
<N as FromStr>::Err: Debug,

Returns an iterator over the converted items.

Panic

Panics when an error occures.